Splitted components-business context content
